pyproject.toml
[project]
name = "postgres-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
